refactor: consolidate components and entities
