+++ src/cala/models/entity.py
@@ -6,6 +6,19 @@ from pydantic import BaseModel, Field
 from cala.models.axis import Coord, Dim, Dims
 
 
+class Component(Enum):
+    """Enumeration of possible component types in the imaging data.
+
+    Attributes:
+        NEURON: Represents neuronal components.
+        BACKGROUND: Represents background components (non-neuronal signals).
+    """
+
+    NEURON = "neuron"
+    BACKGROUND = "background"
+    UNKNOWN = "unknown"
+
+
 class Entity(BaseModel):
     """
     A base entity describable with an xarray dataarray.
