+++ tests/fixtures/simulation.py
@@ -8,7 +8,7 @@ import xarray as xr
 from scipy.ndimage import gaussian_filter
 from scipy.stats.qmc import PoissonDisk
 
-from cala.models.components import Component
+from cala.models.entity import Component
 
 
 @dataclass
